+++ powerflex/provider/device_datasource_test.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package powerflex
+package provider
 
 import (
 	"regexp"
@@ -50,7 +50,7 @@ func TestAccDeviceDatasource(t *testing.T) {
 			{
 				Config: ProviderConfigForTesting + deviceDataWithID,
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("data.powerflex_device.dev4", "id", "c7fc68a200000000"),
+					resource.TestCheckResourceAttr("data.powerflex_device.dev4", "id", "37ef393e00030000"),
 				),
 			},
 			{
@@ -75,7 +75,7 @@ func TestAccDeviceDatasource(t *testing.T) {
 			{
 				Config: ProviderConfigForTesting + deviceDataWithSdsID,
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("data.powerflex_device.dev8", "sds_id", "db2c37000000000"),
+					resource.TestCheckResourceAttr("data.powerflex_device.dev8", "sds_id", "0db7306f00000003"),
 				),
 			},
 			{
@@ -149,7 +149,7 @@ data "powerflex_device" "dev3" {
 
 var deviceDataWithID = `
 data "powerflex_device" "dev4" {
-	id = "c7fc68a200000000"
+	id = "37ef393e00030000"
 }
 `
 var deviceDataWithStoragePoolName = `
@@ -173,7 +173,7 @@ data "powerflex_device" "dev7" {
 
 var deviceDataWithSdsID = `
 data "powerflex_device" "dev8" {
-	sds_id = "db2c37000000000"
+	sds_id = "0db7306f00000003"
 }
 `
