+++ powerflex/provider/sdc_resource.go
@@ -15,27 +15,22 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package powerflex
+package provider
 
 import (
 	"context"
-	"encoding/csv"
-	"encoding/json"
 	"fmt"
-	"os"
 	"strconv"
 	"strings"
-	"terraform-provider-powerflex/helper"
-	"time"
+	"terraform-provider-powerflex/powerflex/helper"
+	"terraform-provider-powerflex/powerflex/models"
 
 	"github.com/dell/goscaleio"
 	goscaleio_types "github.com/dell/goscaleio/types/v1"
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 )
 
@@ -93,7 +88,7 @@ func (r *sdcResource) Configure(_ context.Context, req resource.ConfigureRequest
 func (r *sdcResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	tflog.Debug(ctx, "[POWERFLEX] Create")
 
-	var plan sdcResourceModel
+	var plan models.SdcResourceModel
 
 	diags := req.Plan.Get(ctx, &plan)
 
@@ -102,11 +97,11 @@ func (r *sdcResource) Create(ctx context.Context, req resource.CreateRequest, re
 		return
 	}
 
-	sdcDetailList := []SDCDetailDataModel{}
+	sdcDetailList := []models.SDCDetailDataModel{}
 	diags = plan.SDCDetails.ElementsAs(ctx, &sdcDetailList, true)
 	resp.Diagnostics.Append(diags...)
 
-	system, err := getFirstSystem(r.client)
+	system, err := helper.GetFirstSystem(r.client)
 
 	if err != nil {
 		resp.Diagnostics.AddError(
@@ -116,7 +111,7 @@ func (r *sdcResource) Create(ctx context.Context, req resource.CreateRequest, re
 		return
 	}
 
-	var chnagedSDCs []SDCDetailDataModel
+	var chnagedSDCs []models.SDCDetailDataModel
 
 	if plan.Name.ValueString() != "" && plan.ID.ValueString() != "" {
 
@@ -141,11 +136,11 @@ func (r *sdcResource) Create(ctx context.Context, req resource.CreateRequest, re
 			return
 		}
 
-		changedSDCDetail := getSDCState(*finalSDC.Sdc, SDCDetailDataModel{})
+		changedSDCDetail := helper.GetSDCState(*finalSDC.Sdc, models.SDCDetailDataModel{})
 
 		chnagedSDCs = append(chnagedSDCs, changedSDCDetail)
 
-		data, dgs := updateState(chnagedSDCs, plan)
+		data, dgs := helper.UpdateState(chnagedSDCs, plan)
 		resp.Diagnostics.Append(dgs...)
 
 		diags = resp.State.Set(ctx, data)
@@ -160,7 +155,7 @@ func (r *sdcResource) Create(ctx context.Context, req resource.CreateRequest, re
 
 		resp.Diagnostics.Append(r.UpdateSDCNamdPerfProfileOperations(ctx, sdcDetailList, system, &chnagedSDCs)...)
 
-		data, dgs := updateState(chnagedSDCs, plan)
+		data, dgs := helper.UpdateState(chnagedSDCs, plan)
 		resp.Diagnostics.Append(dgs...)
 
 		diags = resp.State.Set(ctx, data)
@@ -175,18 +170,18 @@ func (r *sdcResource) Create(ctx context.Context, req resource.CreateRequest, re
 // Read - function to Read for SDC resource.
 func (r *sdcResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	tflog.Debug(ctx, "[POWERFLEX] Read")
-	var state sdcResourceModel
+	var state models.SdcResourceModel
 	diags := req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	sdcDetailList := []SDCDetailDataModel{}
+	sdcDetailList := []models.SDCDetailDataModel{}
 	diags = state.SDCDetails.ElementsAs(ctx, &sdcDetailList, true)
 	resp.Diagnostics.Append(diags...)
 
-	system, err := getFirstSystem(r.client)
+	system, err := helper.GetFirstSystem(r.client)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error in getting system instance on the PowerFlex cluster",
@@ -195,7 +190,7 @@ func (r *sdcResource) Read(ctx context.Context, req resource.ReadRequest, resp *
 		return
 	}
 
-	var chnagedSDCs []SDCDetailDataModel
+	var chnagedSDCs []models.SDCDetailDataModel
 
 	//For handling the import case
 	if state.ID.ValueString() != "" && state.ID.ValueString() != "placeholder" && (state.Name.ValueString() == "" || state.Name.IsNull()) {
@@ -212,7 +207,7 @@ func (r *sdcResource) Read(ctx context.Context, req resource.ReadRequest, resp *
 			}
 
 			if sdcData != nil {
-				changedSDCDetail := getSDCState(*sdcData.Sdc, SDCDetailDataModel{})
+				changedSDCDetail := helper.GetSDCState(*sdcData.Sdc, models.SDCDetailDataModel{})
 
 				chnagedSDCs = append(chnagedSDCs, changedSDCDetail)
 			}
@@ -230,7 +225,7 @@ func (r *sdcResource) Read(ctx context.Context, req resource.ReadRequest, resp *
 			return
 		}
 
-		changedSDCDetail := getSDCState(*singleSdc.Sdc, SDCDetailDataModel{})
+		changedSDCDetail := helper.GetSDCState(*singleSdc.Sdc, models.SDCDetailDataModel{})
 
 		chnagedSDCs = append(chnagedSDCs, changedSDCDetail)
 	} else if len(sdcDetailList) > 0 {
@@ -272,12 +267,12 @@ func (r *sdcResource) Read(ctx context.Context, req resource.ReadRequest, resp *
 				}
 
 				if sdcData != nil {
-					changedSDCDetail := getSDCState(*sdcData.Sdc, sdc)
+					changedSDCDetail := helper.GetSDCState(*sdcData.Sdc, sdc)
 
 					chnagedSDCs = append(chnagedSDCs, changedSDCDetail)
 				}
 			} else {
-				changedSDCDetail := getSDCState(*&goscaleio_types.Sdc{}, sdc)
+				changedSDCDetail := helper.GetSDCState(*&goscaleio_types.Sdc{}, sdc)
 
 				chnagedSDCs = append(chnagedSDCs, changedSDCDetail)
 			}
@@ -288,7 +283,7 @@ func (r *sdcResource) Read(ctx context.Context, req resource.ReadRequest, resp *
 		return
 	}
 
-	data, dgs := updateState(chnagedSDCs, state)
+	data, dgs := helper.UpdateState(chnagedSDCs, state)
 	resp.Diagnostics.Append(dgs...)
 
 	diags = resp.State.Set(ctx, data)
@@ -302,27 +297,27 @@ func (r *sdcResource) Read(ctx context.Context, req resource.ReadRequest, resp *
 func (r *sdcResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
 	tflog.Debug(ctx, "[POWERFLEX] Update")
 	// Retrieve values from plan
-	var plan sdcResourceModel
+	var plan models.SdcResourceModel
 	diags := req.Plan.Get(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	var state sdcResourceModel
+	var state models.SdcResourceModel
 	diags = req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	system, err := getFirstSystem(r.client)
+	system, err := helper.GetFirstSystem(r.client)
 
-	planSdcDetailList := []SDCDetailDataModel{}
+	planSdcDetailList := []models.SDCDetailDataModel{}
 	diags = plan.SDCDetails.ElementsAs(ctx, &planSdcDetailList, true)
 	resp.Diagnostics.Append(diags...)
 
-	stateSdcDetailList := []SDCDetailDataModel{}
+	stateSdcDetailList := []models.SDCDetailDataModel{}
 	diags = state.SDCDetails.ElementsAs(ctx, &stateSdcDetailList, true)
 	resp.Diagnostics.Append(diags...)
 
@@ -334,9 +329,9 @@ func (r *sdcResource) Update(ctx context.Context, req resource.UpdateRequest, re
 		return
 	}
 
-	var chnagedSDCs []SDCDetailDataModel
+	var chnagedSDCs []models.SDCDetailDataModel
 
-	deletedSDC := findDeletedSDC(stateSdcDetailList, planSdcDetailList)
+	deletedSDC := helper.FindDeletedSDC(stateSdcDetailList, planSdcDetailList)
 
 	if !(plan.Name.ValueString() != "" && plan.ID.ValueString() != "") {
 		if len(deletedSDC) > 0 {
@@ -381,11 +376,11 @@ func (r *sdcResource) Update(ctx context.Context, req resource.UpdateRequest, re
 			return
 		}
 
-		changedSDCDetail := getSDCState(*finalSDC.Sdc, SDCDetailDataModel{})
+		changedSDCDetail := helper.GetSDCState(*finalSDC.Sdc, models.SDCDetailDataModel{})
 
 		chnagedSDCs = append(chnagedSDCs, changedSDCDetail)
 
-		data, dgs := updateState(chnagedSDCs, plan)
+		data, dgs := helper.UpdateState(chnagedSDCs, plan)
 		resp.Diagnostics.Append(dgs...)
 
 		diags = resp.State.Set(ctx, data)
@@ -405,18 +400,18 @@ func (r *sdcResource) Update(ctx context.Context, req resource.UpdateRequest, re
 					sdcData, _ := system.FindSdc("SdcIP", sdc.IP.ValueString())
 
 					if sdcData != nil {
-						changedSDCDetail := getSDCState(*sdcData.Sdc, sdc)
+						changedSDCDetail := helper.GetSDCState(*sdcData.Sdc, sdc)
 
 						chnagedSDCs = append(chnagedSDCs, changedSDCDetail)
 					}
 				} else {
-					changedSDCDetail := getSDCState(*&goscaleio_types.Sdc{}, sdc)
+					changedSDCDetail := helper.GetSDCState(*&goscaleio_types.Sdc{}, sdc)
 
 					chnagedSDCs = append(chnagedSDCs, changedSDCDetail)
 				}
 			}
 
-			data, dgs := updateState(chnagedSDCs, plan)
+			data, dgs := helper.UpdateState(chnagedSDCs, plan)
 			resp.Diagnostics.Append(dgs...)
 
 			diags = resp.State.Set(ctx, data)
@@ -427,7 +422,7 @@ func (r *sdcResource) Update(ctx context.Context, req resource.UpdateRequest, re
 
 		resp.Diagnostics.Append(r.UpdateSDCNamdPerfProfileOperations(ctx, planSdcDetailList, system, &chnagedSDCs)...)
 
-		data, dgs := updateState(chnagedSDCs, plan)
+		data, dgs := helper.UpdateState(chnagedSDCs, plan)
 		resp.Diagnostics.Append(dgs...)
 
 		diags = resp.State.Set(ctx, data)
@@ -446,18 +441,18 @@ func (r *sdcResource) Update(ctx context.Context, req resource.UpdateRequest, re
 // Delete - function to Delete for SDC resource.
 func (r *sdcResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 	tflog.Debug(ctx, "[POWERFLEX] Delete")
-	var state sdcResourceModel
+	var state models.SdcResourceModel
 	diags := req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	sdcDetailList := []SDCDetailDataModel{}
+	sdcDetailList := []models.SDCDetailDataModel{}
 	diags = state.SDCDetails.ElementsAs(ctx, &sdcDetailList, true)
 	resp.Diagnostics.Append(diags...)
 
-	system, err := getFirstSystem(r.client)
+	system, err := helper.GetFirstSystem(r.client)
 
 	if err != nil {
 		resp.Diagnostics.AddError(
@@ -492,157 +487,11 @@ func (r *sdcResource) ImportState(ctx context.Context, req resource.ImportStateR
 	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
 }
 
-// findDeletedSDC function to find deleted SDC Details in Plan
-func findDeletedSDC(state, plan []SDCDetailDataModel) []SDCDetailDataModel {
-	difference := []SDCDetailDataModel{}
-
-	for _, obj1 := range state {
-		found := false
-		for _, obj2 := range plan {
-			if obj2.IP.ValueString() != "" && obj1.IP == obj2.IP {
-				found = true
-				break
-			} else if obj2.SDCID.ValueString() != "" && obj1.SDCID == obj2.SDCID {
-				found = true
-				break
-			} else if obj2.SDCName.ValueString() != "" && obj1.SDCName == obj2.SDCName {
-				found = true
-				break
-			}
-		}
-		if !found {
-			difference = append(difference, obj1)
-		}
-	}
-
-	return difference
-}
-
-// getSDCDetailType returns the SDC Detail type
-func getSDCDetailType() map[string]attr.Type {
-	return map[string]attr.Type{
-		"sdc_id":               types.StringType,
-		"ip":                   types.StringType,
-		"username":             types.StringType,
-		"password":             types.StringType,
-		"operating_system":     types.StringType,
-		"is_mdm_or_tb":         types.StringType,
-		"is_sdc":               types.StringType,
-		"performance_profile":  types.StringType,
-		"name":                 types.StringType,
-		"system_id":            types.StringType,
-		"sdc_approved":         types.BoolType,
-		"on_vmware":            types.BoolType,
-		"sdc_guid":             types.StringType,
-		"mdm_connection_state": types.StringType,
-	}
-}
-
-// getSDCDetailValue returns the SDC Detail model object value
-func getSDCDetailValue(sdc SDCDetailDataModel) (basetypes.ObjectValue, diag.Diagnostics) {
-	return types.ObjectValue(getSDCDetailType(), map[string]attr.Value{
-		"sdc_id":               types.StringValue(sdc.SDCID.ValueString()),
-		"ip":                   types.StringValue(sdc.IP.ValueString()),
-		"username":             types.StringValue(sdc.UserName.ValueString()),
-		"password":             types.StringValue(sdc.Password.ValueString()),
-		"operating_system":     types.StringValue(sdc.OperatingSystem.ValueString()),
-		"is_mdm_or_tb":         types.StringValue(sdc.IsMdmOrTb.ValueString()),
-		"is_sdc":               types.StringValue(sdc.IsSdc.ValueString()),
-		"performance_profile":  types.StringValue(sdc.PerformanceProfile.ValueString()),
-		"name":                 types.StringValue(sdc.SDCName.ValueString()),
-		"system_id":            types.StringValue(sdc.SystemID.ValueString()),
-		"sdc_approved":         types.BoolValue(sdc.SdcApproved.ValueBool()),
-		"on_vmware":            types.BoolValue(sdc.OnVMWare.ValueBool()),
-		"sdc_guid":             types.StringValue(sdc.SdcGUID.ValueString()),
-		"mdm_connection_state": types.StringValue(sdc.MdmConnectionState.ValueString()),
-	})
-}
-
-// updateState - function to update state file for SDC resource.
-func updateState(sdcs []SDCDetailDataModel, plan sdcResourceModel) (sdcResourceModel, diag.Diagnostics) {
-	state := plan
-	var diags diag.Diagnostics
-
-	SDCAttrTypes := getSDCDetailType()
-
-	SDCElemType := types.ObjectType{
-		AttrTypes: SDCAttrTypes,
-	}
-
-	objectSDCs := []attr.Value{}
-	for _, sdc := range sdcs {
-		objVal, dgs := getSDCDetailValue(sdc)
-		diags = append(diags, dgs...)
-		objectSDCs = append(objectSDCs, objVal)
-	}
-	setSdcs, dgs := types.ListValue(SDCElemType, objectSDCs)
-	diags = append(diags, dgs...)
-
-	state.SDCDetails = setSdcs
-
-	if plan.ID.ValueString() != "" && len(strings.Split(plan.ID.ValueString(), ",")) == 1 {
-		state.ID = plan.ID
-	} else {
-		state.ID = types.StringValue("placeholder")
-	}
-
-	return state, diags
-}
-
-// GetMDMIP function is used for fetch MDM IP from cluster details
-func GetMDMIP(ctx context.Context, sdcDetails []SDCDetailDataModel) (string, error) {
-	var mdmIP string
-
-	for _, item := range sdcDetails {
-		if strings.EqualFold(item.IsMdmOrTb.ValueString(), "Primary") {
-			mdmIP = item.IP.ValueString()
-			return mdmIP, nil
-		}
-	}
-	return mdmIP, nil
-}
-
-// CheckForExpansion function is used for check for expansion
-func CheckForExpansion(model []SDCDetailDataModel) bool {
-	performaneChangeSdc := false
-
-	for _, item := range model {
-		if item.Password.ValueString() != "" && strings.EqualFold(item.IsSdc.ValueString(), "Yes") {
-			performaneChangeSdc = true
-			break
-		}
-	}
-	return performaneChangeSdc
-}
-
-// ResetInstallerQueue function for the Abort, Clear and Move To Idle Execution
-func ResetInstallerQueue(gatewayClient *goscaleio.GatewayClient) error {
-
-	_, err := gatewayClient.AbortOperation()
-
-	if err != nil {
-		return fmt.Errorf("Error while Aborting Operation is %s", err.Error())
-	}
-	_, err = gatewayClient.ClearQueueCommand()
-
-	if err != nil {
-		return fmt.Errorf("Error while Clearing Queue is %s", err.Error())
-	}
-
-	_, err = gatewayClient.MoveToIdlePhase()
-
-	if err != nil {
-		return fmt.Errorf("Error while Move to Ideal Phase is %s", err.Error())
-	}
-
-	return nil
-}
-
 // SDCExpansionOperations function for the SDC Expansion Operation Like ParseCSV, Validate MDM and Installation
-func (r *sdcResource) SDCExpansionOperations(ctx context.Context, plan sdcResourceModel, system *goscaleio.System, sdcDetails []SDCDetailDataModel) (dia diag.Diagnostics) {
+func (r *sdcResource) SDCExpansionOperations(ctx context.Context, plan models.SdcResourceModel, system *goscaleio.System, sdcDetails []models.SDCDetailDataModel) (dia diag.Diagnostics) {
 
-	if CheckForExpansion(sdcDetails) {
-		parsecsvRespose, parseCSVError := ParseCSVOperation(ctx, sdcDetails, r.gatewayClient)
+	if helper.CheckForExpansion(sdcDetails) {
+		parsecsvRespose, parseCSVError := helper.ParseCSVOperation(ctx, sdcDetails, r.gatewayClient)
 
 		if parseCSVError != nil {
 			dia.AddError(
@@ -653,7 +502,7 @@ func (r *sdcResource) SDCExpansionOperations(ctx context.Context, plan sdcResour
 		}
 
 		// to make gateway available for installation
-		queueOperationError := ResetInstallerQueue(r.gatewayClient)
+		queueOperationError := helper.ResetInstallerQueue(r.gatewayClient)
 		if queueOperationError != nil {
 			dia.AddError(
 				"Error Clearing Queue",
@@ -664,7 +513,7 @@ func (r *sdcResource) SDCExpansionOperations(ctx context.Context, plan sdcResour
 
 		tflog.Info(ctx, "Gateway Installer changed to idle phase before initiating process")
 
-		mdmIP, mdmIPError := GetMDMIP(ctx, sdcDetails)
+		mdmIP, mdmIPError := helper.GetMDMIP(ctx, sdcDetails)
 		if mdmIPError != nil {
 			dia.AddError(
 				"Error while Getting MDM IP",
@@ -676,7 +525,7 @@ func (r *sdcResource) SDCExpansionOperations(ctx context.Context, plan sdcResour
 		tflog.Info(ctx, "CSV File parsed ssucessfully")
 
 		// Vaidate the MDM credentials
-		validateMDMResponse, validateMDMError := ValidateMDMOperation(ctx, plan, r.gatewayClient, mdmIP)
+		validateMDMResponse, validateMDMError := helper.ValidateMDMOperation(ctx, plan, r.gatewayClient, mdmIP)
 		if validateMDMError != nil {
 			dia.AddError(
 				"Error While Validating MDM Details",
@@ -689,8 +538,8 @@ func (r *sdcResource) SDCExpansionOperations(ctx context.Context, plan sdcResour
 
 			tflog.Info(ctx, "MDM Details validated successfully")
 
-			if !CheckForNewSDCIPs(strings.Split(parsecsvRespose.Message, ","), strings.Split(validateMDMResponse.Data, ",")) {
-				installationError := InstallationOperations(ctx, plan, r.gatewayClient, parsecsvRespose)
+			if !helper.CheckForNewSDCIPs(strings.Split(parsecsvRespose.Message, ","), strings.Split(validateMDMResponse.Data, ",")) {
+				installationError := helper.InstallationOperations(ctx, plan, r.gatewayClient, parsecsvRespose)
 
 				if installationError != nil {
 					dia.AddError(
@@ -713,7 +562,7 @@ func (r *sdcResource) SDCExpansionOperations(ctx context.Context, plan sdcResour
 }
 
 // UpdateSDCNamdPerfProfileOperations function for Update Name and Performance Profile of SDC
-func (r *sdcResource) UpdateSDCNamdPerfProfileOperations(ctx context.Context, sdcDetailList []SDCDetailDataModel, system *goscaleio.System, chnagedSDCs *[]SDCDetailDataModel) (dia diag.Diagnostics) {
+func (r *sdcResource) UpdateSDCNamdPerfProfileOperations(ctx context.Context, sdcDetailList []models.SDCDetailDataModel, system *goscaleio.System, chnagedSDCs *[]models.SDCDetailDataModel) (dia diag.Diagnostics) {
 
 	for _, sdc := range sdcDetailList {
 
@@ -728,7 +577,7 @@ func (r *sdcResource) UpdateSDCNamdPerfProfileOperations(ctx context.Context, sd
 			}
 
 			if sdcData != nil {
-				changedSDCDetail := getSDCState(*sdcData.Sdc, sdc)
+				changedSDCDetail := helper.GetSDCState(*sdcData.Sdc, sdc)
 
 				*chnagedSDCs = append(*chnagedSDCs, changedSDCDetail)
 			}
@@ -761,7 +610,7 @@ func (r *sdcResource) UpdateSDCNamdPerfProfileOperations(ctx context.Context, sd
 
 			if sdc.SDCName.ValueString() != "" && sdc.SDCID.ValueString() != "" {
 
-				nameExist, _ := checkForSDCName(system, sdc)
+				nameExist, _ := helper.CheckForSDCName(system, sdc)
 
 				if !nameExist {
 					nameChng, err := system.ChangeSdcName(sdc.SDCID.ValueString(), sdc.SDCName.ValueString())
@@ -800,13 +649,13 @@ func (r *sdcResource) UpdateSDCNamdPerfProfileOperations(ctx context.Context, sd
 			}
 
 			if finalSDC != nil {
-				changedSDCDetail := getSDCState(*finalSDC.Sdc, sdc)
+				changedSDCDetail := helper.GetSDCState(*finalSDC.Sdc, sdc)
 
 				*chnagedSDCs = append(*chnagedSDCs, changedSDCDetail)
 			}
 		} else if strings.EqualFold(sdc.IsSdc.ValueString(), "No") {
 
-			changedSDCDetail := getSDCState(goscaleio_types.Sdc{}, sdc)
+			changedSDCDetail := helper.GetSDCState(goscaleio_types.Sdc{}, sdc)
 
 			*chnagedSDCs = append(*chnagedSDCs, changedSDCDetail)
 
@@ -814,250 +663,4 @@ func (r *sdcResource) UpdateSDCNamdPerfProfileOperations(ctx context.Context, sd
 	}
 
 	return
-}
-
-// ParseCSVOperation function for Handling Parsing CSV Operation
-func ParseCSVOperation(ctx context.Context, sdcDetails []SDCDetailDataModel, gatewayClient *goscaleio.GatewayClient) (*goscaleio_types.GatewayResponse, error) {
-
-	var parseCSVResponse goscaleio_types.GatewayResponse
-
-	//Create a csv file from the input given by the user
-	mydir, err := os.Getwd()
-	if err != nil {
-		return &parseCSVResponse, fmt.Errorf("Error While Reading Current Directory is %s", err.Error())
-	}
-	// Create a csv writer
-	file, err := os.Create(mydir + "/Minimal.csv")
-	if err != nil {
-		return &parseCSVResponse, fmt.Errorf("Error While Creating Temp CSV is %s", err.Error())
-	}
-	defer file.Close()
-	writer := csv.NewWriter(file)
-
-	// Write the header row
-	header := []string{"IPs", "Username", "Password", "Operating System", "Is MDM/TB", "Is SDC", "perfProfileForSDC"}
-	err = writer.Write(header)
-	if err != nil {
-		return &parseCSVResponse, fmt.Errorf("Error While Writing Temp CSV is %s", err.Error())
-	}
-
-	var sdcIPs []string
-
-	for _, item := range sdcDetails {
-
-		if item.Password.ValueString() != "" {
-			// Add mapped SDC
-			csvStruct := CsvRow{
-				IP:              item.IP.ValueString(),
-				UserName:        item.UserName.ValueString(),
-				Password:        item.Password.ValueString(),
-				IsMdmOrTb:       item.IsMdmOrTb.ValueString(),
-				OperatingSystem: item.OperatingSystem.ValueString(),
-				IsSdc:           item.IsSdc.ValueString(),
-			}
-
-			if strings.EqualFold(csvStruct.IsSdc, "Yes") {
-				sdcIPs = append(sdcIPs, csvStruct.IP)
-			}
-
-			if strings.EqualFold(item.PerformanceProfile.ValueString(), "HighPerformance") {
-				csvStruct.PerformanceProfile = "High"
-			}
-
-			//Write the data row
-			data := []string{csvStruct.IP, csvStruct.UserName, csvStruct.Password, csvStruct.OperatingSystem, csvStruct.IsMdmOrTb, csvStruct.IsSdc, csvStruct.PerformanceProfile} //, csvStruct.SDCName
-			err = writer.Write(data)
-			if err != nil {
-				return &parseCSVResponse, fmt.Errorf("Error While Creating Temp CSV File is %s", err.Error())
-			}
-		}
-
-	}
-	writer.Flush()
-
-	parsecsvRespose, parseCSVError := gatewayClient.ParseCSV(mydir + "/Minimal.csv")
-
-	deletCSVError := os.Remove(mydir + "/Minimal.csv")
-	if deletCSVError != nil {
-		return &parseCSVResponse, fmt.Errorf("Error While Deleting Temp CSV File is %s", deletCSVError.Error())
-	}
-
-	if parseCSVError != nil {
-		return &parseCSVResponse, fmt.Errorf("%s", parseCSVError.Error())
-	}
-
-	parsecsvRespose.Message = strings.Join(sdcIPs, ",")
-
-	if parsecsvRespose.StatusCode != 200 {
-		return &parseCSVResponse, fmt.Errorf("Meesage : %s, Error Cosde : %s", parsecsvRespose.Message, strconv.Itoa(parsecsvRespose.StatusCode))
-	}
-
-	return parsecsvRespose, nil
-}
-
-// ValidateMDMOperation function for Validate the MDM credentials
-func ValidateMDMOperation(ctx context.Context, model sdcResourceModel, gatewayClient *goscaleio.GatewayClient, mdmIP string) (*goscaleio_types.GatewayResponse, error) {
-	mapData := map[string]interface{}{
-		"mdmUser":     "admin",
-		"mdmPassword": model.MdmPassword.ValueString(),
-	}
-	mapData["mdmIps"] = []string{mdmIP}
-
-	secureData := map[string]interface{}{
-		"allowNonSecureCommunicationWithMdm": true,
-		"allowNonSecureCommunicationWithLia": true,
-		"disableNonMgmtComponentsAuth":       false,
-	}
-	mapData["securityConfiguration"] = secureData
-	jsonres, _ := json.Marshal(mapData)
-
-	validateMDMResponse, validateMDMError := gatewayClient.ValidateMDMDetails(jsonres)
-	if validateMDMError != nil {
-		return validateMDMResponse, fmt.Errorf("%s", validateMDMError.Error())
-	}
-
-	return validateMDMResponse, nil
-}
-
-// InstallationOperations function for begin instllation process
-func InstallationOperations(ctx context.Context, model sdcResourceModel, gatewayClient *goscaleio.GatewayClient, parsecsvRespose *goscaleio_types.GatewayResponse) error {
-
-	beginInstallationResponse, installationError := gatewayClient.BeginInstallation(parsecsvRespose.Data, "admin", model.MdmPassword.ValueString(), model.LiaPassword.ValueString(), true)
-
-	if installationError != nil {
-		return fmt.Errorf("Error while begin installation is %s", installationError.Error())
-	}
-
-	if beginInstallationResponse.StatusCode == 200 {
-		currentPhase := "query"
-		couterForStopExecution := 0
-
-		tflog.Info(ctx, "Gateway Installation Begin, Current Phase - Query")
-
-		for couterForStopExecution <= 5 {
-
-			time.Sleep(1 * time.Minute)
-
-			checkForPhaseCompleted, _ := gatewayClient.CheckForCompletionQueueCommands(currentPhase)
-
-			if checkForPhaseCompleted.Data == "Completed" {
-				couterForStopExecution = 0
-
-				if currentPhase != "configure" {
-					moveToNextPhaseResponse, err := gatewayClient.MoveToNextPhase()
-
-					if err != nil {
-						return fmt.Errorf("Error while moving to next phase is %s", err.Error())
-					}
-
-					if moveToNextPhaseResponse.StatusCode == 200 {
-						if currentPhase == "query" {
-							currentPhase = "upload"
-							tflog.Info(ctx, "Gateway Installation phase changed to Upload")
-						} else if currentPhase == "upload" {
-							currentPhase = "install"
-							tflog.Info(ctx, "Gateway Installation phase changed to Install")
-						} else if currentPhase == "install" {
-							currentPhase = "configure"
-							tflog.Info(ctx, "Gateway Installation phase changed to Configure")
-						}
-					} else {
-						return fmt.Errorf("Messsage: %s, Error Code: %s", moveToNextPhaseResponse.Message, strconv.Itoa(moveToNextPhaseResponse.StatusCode))
-					}
-				} else {
-					// to make gateway available for installation
-					queueOperationError := ResetInstallerQueue(gatewayClient)
-					if queueOperationError != nil {
-						return fmt.Errorf("Error Clearing Queue During Installation is %s", queueOperationError.Error())
-					}
-
-					couterForStopExecution = 10
-
-					return nil
-				}
-
-			} else if checkForPhaseCompleted.Data == "Running" {
-				couterForStopExecution++
-
-				tflog.Info(ctx, "Gateway Installation operations are still running")
-
-				if couterForStopExecution == 5 {
-					// to make gateway available for installation
-					queueOperationError := ResetInstallerQueue(gatewayClient)
-					if queueOperationError != nil {
-						return fmt.Errorf("Error Clearing Queue During Installation is %s", queueOperationError.Error())
-					}
-
-					return fmt.Errorf("Time Out,Some Operations of Installer running from since long")
-				}
-
-			} else {
-				return fmt.Errorf("Error During Installation is %s", checkForPhaseCompleted.Message)
-			}
-		}
-	} else {
-		return fmt.Errorf("Message: %s, Error Code: %s", beginInstallationResponse.Message, strconv.Itoa(beginInstallationResponse.StatusCode))
-	}
-
-	return nil
-}
-
-// CheckForNewSDCIPs function to check SDC Alredy Installed or not
-func CheckForNewSDCIPs(newSDCIPS []string, installedSDCIPs []string) bool {
-	checkset := make(map[string]bool)
-	for _, element := range newSDCIPS {
-		checkset[element] = true
-	}
-	for _, value := range installedSDCIPs {
-		if checkset[value] {
-			delete(checkset, value)
-		}
-	}
-	return len(checkset) == 0 //this implies that set is subset of superset
-}
-
-// getSDCState - function to return sdc result from goscaleio.
-func getSDCState(sdc goscaleio_types.Sdc, model SDCDetailDataModel) (response SDCDetailDataModel) {
-
-	if sdc.ID != "" {
-		model.SDCID = types.StringValue(sdc.ID)
-
-		model.SDCName = types.StringValue(sdc.Name)
-
-		if sdc.SdcGUID != "" {
-			model.SdcGUID = types.StringValue(sdc.SdcGUID)
-		}
-
-		model.SdcApproved = types.BoolValue(sdc.SdcApproved)
-
-		model.OnVMWare = types.BoolValue(sdc.OnVMWare)
-
-		if sdc.SystemID != "" {
-			model.SystemID = types.StringValue(sdc.SystemID)
-		}
-
-		model.PerformanceProfile = types.StringValue(sdc.PerfProfile)
-
-		if sdc.SdcIP != "" {
-			model.IP = types.StringValue(sdc.SdcIP)
-		}
-
-		if sdc.MdmConnectionState != "" {
-			model.MdmConnectionState = types.StringValue(sdc.MdmConnectionState)
-		}
-	}
-
-	return model
-}
-
-// checkForSDCName - check for the SDC Name already exist or not
-func checkForSDCName(system *goscaleio.System, sdcDetail SDCDetailDataModel) (bool, error) {
-
-	sdcData, err := system.FindSdc("Name", sdcDetail.SDCName.ValueString())
-
-	if err == nil && sdcData.Sdc.ID == sdcDetail.SDCID.ValueString() {
-		return true, fmt.Errorf("SDC Name:%s already exist", sdcDetail.SDCName.ValueString())
-	}
-
-	return false, nil
 }
