+++ powerflex/provider/sdc_resource_test.go
@@ -15,15 +15,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package powerflex
+package provider
 
 import (
+	"fmt"
 	"os"
 	"regexp"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
 // TestAccSDCResource tests the SDC Expansion Operation
@@ -50,53 +53,172 @@ func TestAccSDCResource(t *testing.T) {
 			//Create with Packages
 			{
 				Config: ProviderConfigForTesting + packageTest + SDCConfig2,
-				Check: resource.TestCheckTypeSetElemNestedAttrs("powerflex_sdc.test", "sdc_details.*", map[string]string{
-					"ip": GatewayDataPoints.tbIP,
-				}),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("powerflex_sdc.test", "sdc_details.2.ip", GatewayDataPoints.tbIP),
+				),
 			},
 			//Update
 			{
 				Config: ProviderConfigForTesting + packageTest + SDCConfigUpdate,
-				Check: resource.TestCheckTypeSetElemNestedAttrs("powerflex_sdc.test", "sdc_details.*", map[string]string{
-					"ip": GatewayDataPoints.sdcServerIP,
-				}),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("powerflex_sdc.test", "sdc_details.3.ip", GatewayDataPoints.sdcServerIP),
+				),
 			},
 			//Reaname
 			{
 				Config: ProviderConfigForTesting + packageTest + SDCConfigRename,
-				Check: resource.TestCheckTypeSetElemNestedAttrs("powerflex_sdc.test", "sdc_details.*", map[string]string{
-					"name": time.Now().Weekday().String(),
-					"ip":   GatewayDataPoints.secondaryMDMIP,
-				}),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("powerflex_sdc.test", "sdc_details.1.ip", GatewayDataPoints.secondaryMDMIP),
+					resource.TestCheckResourceAttr("powerflex_sdc.test", "sdc_details.1.name", time.Now().Weekday().String()),
+				),
 			},
-			//Perormance Profile
+			//Performance Profile
 			{
 				Config: ProviderConfigForTesting + packageTest + SDCConfigPerProfile,
-				Check: resource.TestCheckTypeSetElemNestedAttrs("powerflex_sdc.test", "sdc_details.*", map[string]string{
-					"ip":                  GatewayDataPoints.secondaryMDMIP,
-					"performance_profile": "Compact",
-				}),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("powerflex_sdc.test", "sdc_details.1.ip", GatewayDataPoints.secondaryMDMIP),
+					resource.TestCheckResourceAttr("powerflex_sdc.test", "sdc_details.1.performance_profile", "HighPerformance"),
+				),
 			},
 		},
 	})
 }
 
-func TestAccSDCManagerResourceNegative(t *testing.T) {
+func generateDynamicConfig(baseConfig string, sdcID string, name string) string {
+	return fmt.Sprintf("%s\nresource \"powerflex_sdc\" \"test\" {\n  name = \"%s\"\n  id = \"%s\"\n}", baseConfig, name, sdcID)
+}
+
+func TestAccSDCResourceSingleNameOperation(t *testing.T) {
+
+	var sdcID string // Variable to store the SDC IP
+
 	resource.Test(t, resource.TestCase{
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: ProviderConfigForTesting + SDCConfigChangeName,
-				Check: resource.TestCheckTypeSetElemNestedAttrs("powerflex_sdc.name", "sdc_details.*", map[string]string{
-					"name": time.Now().Weekday().String() + "1",
-				}),
+				Config: ProviderConfigForTesting + TestSdcDataSourceByName,
+
+				Check: resource.ComposeAggregateTestCheckFunc(
+
+					// Store the SDC ID in the variable
+					func(state *terraform.State) error {
+						var index string
+
+						attrs, ok := state.RootModule().Resources["data.powerflex_sdc.selected"].Primary.Attributes, true
+						if !ok {
+							return fmt.Errorf("data resource not found in state")
+						}
+
+						for key, value := range attrs {
+							if strings.Contains(key, "ip") && value == "N/A" {
+								index = strings.Split(key, ".")[1]
+							}
+						}
+
+						sdcID = attrs["sdcs."+index+".id"]
+						return nil
+					},
+				),
+			},
+		}})
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: generateDynamicConfig(ProviderConfigForTesting, sdcID, "terraform_sdc_dummy_2"),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("powerflex_sdc.test", "sdc_details.0.name", "terraform_sdc_dummy_2"),
+				),
 			},
 			{
-				Config: ProviderConfigForTesting + SDCConfigUpdateName,
-				Check: resource.TestCheckTypeSetElemNestedAttrs("powerflex_sdc.name", "sdc_details.*", map[string]string{
-					"name": time.Now().Weekday().String() + "2",
-				}),
+				Config: generateDynamicConfig(ProviderConfigForTesting, sdcID, "terraform_sdc_dummy_3"),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("powerflex_sdc.test", "sdc_details.0.name", "terraform_sdc_dummy_3"),
+				),
 			},
+		}})
+}
+
+func TestAccSDCResourceMultiNameOperation(t *testing.T) {
+
+	var sdcID string // Variable to store the SDC IP
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: ProviderConfigForTesting + TestSdcDataSourceByName,
+
+				Check: resource.ComposeAggregateTestCheckFunc(
+
+					// Store the SDC ID in the variable
+					func(state *terraform.State) error {
+						var index string
+
+						attrs, ok := state.RootModule().Resources["data.powerflex_sdc.selected"].Primary.Attributes, true
+						if !ok {
+							return fmt.Errorf("data resource not found in state")
+						}
+
+						for key, value := range attrs {
+							if strings.Contains(key, "ip") && value == "N/A" {
+								index = strings.Split(key, ".")[1]
+							}
+						}
+
+						sdcID = attrs["sdcs."+index+".id"]
+						return nil
+					},
+				),
+			},
+		}})
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: ProviderConfigForTesting + `resource "powerflex_sdc" "update" {
+					mdm_password = "MDM_Password"
+					lia_password = "LIA_Password"
+					sdc_details = [
+					  {
+						sdc_id = "` + sdcID + `"
+						name   = "terraform_sdc_dummy_2"
+						performance_profile   = "HighPerformance"
+						is_sdc = "Yes"
+					  },
+					]
+				  }`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("powerflex_sdc.update", "sdc_details.0.name", "terraform_sdc_dummy_2"),
+				),
+			},
+			{
+				Config: ProviderConfigForTesting + `resource "powerflex_sdc" "update" {
+					mdm_password = "MDM_Password"
+					lia_password = "LIA_Password"
+					sdc_details = [
+					  {
+						sdc_id = "` + sdcID + `"
+						name   = "terraform_sdc_dummy_3"
+						performance_profile   = "HighPerformance"
+						is_sdc = "Yes"
+					  },
+					]
+				  }`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("powerflex_sdc.update", "sdc_details.0.name", "terraform_sdc_dummy_3"),
+				),
+			},
+		}})
+}
+
+func TestAccSDCResourceNegative(t *testing.T) {
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
 			{
 				Config:      ProviderConfigForTesting + WithoutIP,
 				ExpectError: regexp.MustCompile(`.*Invalid Attribute Value Length.*`),
@@ -135,6 +257,9 @@ resource "powerflex_package" "upload-test" {
 	"/root/powerflex_packages/PowerFlex_3.6.700.103_RHEL_OEL7/EMC-ScaleIO-sdr-3.6-700.103.el7.x86_64.rpm"]
 	}
 `
+var TestSdcDataSourceByName = `data "powerflex_sdc" "selected" {
+}`
+
 var SDCConfigChangeName = `
 resource "powerflex_sdc" "name" {
 	id   = "e3cff47d00000005"
@@ -209,7 +334,7 @@ resource "powerflex_sdc" "test" {
 			password = "` + GatewayDataPoints.serverPassword + `"
 			operating_system = "linux"
 			is_mdm_or_tb = "TB"
-			is_sdc = "Yes"
+			is_sdc = "No"
 			performance_profile = "Compact"
 	    },
 	    {
@@ -240,7 +365,7 @@ resource "powerflex_sdc" "test" {
 			password = "` + GatewayDataPoints.serverPassword + `"
 			operating_system = "linux"
 			is_mdm_or_tb = "Secondary"
-			is_sdc = "Yes"
+			is_sdc = "No"
 			name = "` + time.Now().Weekday().String() + `"
 		},
 		{
@@ -277,14 +402,14 @@ resource "powerflex_sdc" "test" {
 			password = "` + GatewayDataPoints.serverPassword + `"
 			operating_system = "linux"
 			is_mdm_or_tb = "Primary"
-			is_sdc = "Yes"
+			is_sdc = "No"
 		},
 		{
 			ip = "` + GatewayDataPoints.secondaryMDMIP + `"
 			password = "` + GatewayDataPoints.serverPassword + `"
 			operating_system = "linux"
 			is_mdm_or_tb = "Secondary"
-			is_sdc = "Yes"
+			is_sdc = "No"
 			performance_profile = "HighPerformance"
 			name                = "sdc_expansion_test1"
 		},
@@ -293,7 +418,7 @@ resource "powerflex_sdc" "test" {
 			password = "` + GatewayDataPoints.serverPassword + `"
 			operating_system = "linux"
 			is_mdm_or_tb = "TB"
-			is_sdc = "Yes"
+			is_sdc = "No"
 	    },
 	]
 }
@@ -327,14 +452,14 @@ resource "powerflex_sdc" "test" {
 			password = "` + GatewayDataPoints.serverPassword + `"
 			operating_system = "linux"
 			is_mdm_or_tb = "TB"
-			is_sdc = "Yes"
+			is_sdc = "No"
 	    },
 	    {
 			ip = "` + GatewayDataPoints.sdcServerIP + `"
 			password = "` + GatewayDataPoints.serverPassword + `"
 			operating_system = "linux"
 			is_mdm_or_tb = ""
-			is_sdc = "Yes"
+			is_sdc = "No"
 			performance_profile = "HighPerformance"
    		},
 	]
