+++ powerflex/provider/storagepool_data_source_schema.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package powerflex
+package provider
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/listvalidator"
